+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.0] 2024-08-21
+
 ### Added
 
 -   It is now possible to cancel multi-tick coroutines in the middle of execution.
@@ -85,7 +87,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   First implementation of the library.
 
-[Unreleased]: https://github.com/ilyvion/ilyvion-laboratory/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/ilyvion/ilyvion-laboratory/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/ilyvion/ilyvion-laboratory/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/ilyvion/ilyvion-laboratory/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/ilyvion/ilyvion-laboratory/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/ilyvion/ilyvion-laboratory/compare/v0.5.0...v0.6.0